#33: Refactor Character Sheet
Clean up character sheet by moving functionality to helper components.  I’m still on the fence on how I want this to look in the end, but right now I like this a bit better.  Character Sheet looks much more organized.
