+++ Assets/Scripts/View Model/Actor/Stats/DerivedStats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21389814ec0204c59887233a6b7a19c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
